langchain-docs-rag/requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
openai
cohere
rank-bm25
ragas
streamlit
python-dotenv
beautifulsoup4
requests
tiktoken
lxml
